# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Pub/Sub Gateway
  description: HTTPを通してPub/Subが使えるようにするゲートウェイサーバ
  version: ""
host: localhost:8088
basePath: /v1
consumes:
- application/json
produces:
- application/json
paths:
  /gcp/publish/{topic}:
    post:
      tags:
      - gcp
      summary: publish gcp
      description: httpプロトコルを経由してCloud Pub/Subにメッセージを渡すエンドポイント
      operationId: gcp#publish
      parameters:
      - name: topic
        in: path
        description: Cloud Pub/SubのTopic
        required: true
        type: string
        minLength: 1
      - name: PublishRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/GcpPublishRequestBody'
          required:
          - message
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/GcpPublishResponseBody'
      schemes:
      - http
  /healthz/liveness:
    get:
      tags:
      - healthz
      summary: liveness healthz
      description: サーバが生きているかチェックするエンドポイント
      operationId: healthz#liveness
      responses:
        "200":
          description: OK response.
        "500":
          description: Internal Server Error response.
      schemes:
      - http
  /healthz/readiness:
    get:
      tags:
      - healthz
      summary: readiness healthz
      description: サーバが準備完了状態でいるかチェックするエンドポイント
      operationId: healthz#readiness
      responses:
        "200":
          description: OK response.
        "500":
          description: Internal Server Error response.
      schemes:
      - http
definitions:
  GcpPublishRequestBody:
    title: GcpPublishRequestBody
    type: object
    properties:
      message:
        type: string
        description: Cloud Pub/SubのメッセージBody
        example: hz
        minLength: 1
    example:
      message: dd
    required:
    - message
  GcpPublishResponseBody:
    title: 'Mediatype identifier: application/vnd.pubsub-gateway.gcp-publish; view=default'
    type: object
    properties:
      id:
        type: string
        description: message id
        example: Maiores optio ut laboriosam eos totam.
    description: PublishResponseBody result type (default view)
    example:
      id: Nihil necessitatibus debitis sed ut qui aut.
    required:
    - id